api_tests/requirements.txt
allure-pytest==2.10.0
PyHamcrest==2.0.4
pytest==7.1.3
requests==2.28.1
